feat: update card creation and deletion services

+++ src/modules/cards/services/DeleteCardService.ts
@@ -1,12 +1,15 @@
 import { inject, injectable } from 'tsyringe';
 import LibError from '../../../shared/errors/LibError';
 import { ICardRepository } from '../repositories/ICardRepository';
+import { IUserRepository } from '../../users/repositories/IUserRepository';
 
 @injectable()
 export class DeleteCardService {
   constructor(
     @inject('CardRepository')
     private cardRepository: ICardRepository,
+    @inject('UserRepository')
+    private userRepository: IUserRepository,
   ) {}
 
   async execute(id: string): Promise<void> {
@@ -20,6 +23,15 @@ export class DeleteCardService {
       throw new LibError('The card does not exist', 404);
     }
 
+    const user = await this.userRepository.findById(card.user_id);
+
+    if (!user) {
+      throw new LibError('The user does not exist', 404);
+    }
+
+    user.card_ids = user.card_ids?.filter(card_id => card_id !== card.id) || [];
+
+    await this.userRepository.update(user);
     await this.cardRepository.delete(card);
   }
 }
